Bug 1678970 [wpt PR 26376] - [css-grid] Fix up that the width of nested grid container with margin:auto is zero, a=testonly
Automatic update from web-platform-tests
[css-grid] Fix up that the width of nested grid container with margin:auto is zero

the width of nested grid container with margin:auto is zero when their
item has 'max-height'. so the grid item's position is incorrect because
the value of auto-margin computed with 'width' is wrong. Because intrin-
sic width is not updated so they use the old intrinsic width.
This patch is checking the dirty of the intrinsic logical width when
the grid area is changed in |UpdateGridAreaLogicalSize|.

Bug: 1149143
Change-Id: I4d4bf839723f05c60ed91b80148caf7b504d5aa1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2503910
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Javier Fernandez <jfernandez@igalia.com>
Commit-Queue: Hyunjune Kim <hyunjune.kim@samsung.com>
Cr-Commit-Position: refs/heads/master@{#830086}

--

wpt-commits: d21eaee14bab28a8bd57727d41330d0bef170a92
wpt-pr: 26376