Bug 1676559 - Pt 6 - Reduce size of TextureSource / BatchTextures. r=nical
These structs are created and copied around many times during
batching. As PrevPassColor and PrevPassAlpha are removed, the
BatchTextures struct will contain extra fields (such as the clip
mask binding), so it's important to reduce the size of them to
avoid regressiing batching time. This patch reduces the size of
`BatchTextures` from 72 bytes to 24 bytes.

Differential Revision: https://phabricator.services.mozilla.com/D98122