Bug 1605134 - Stop Fallback driver after the first audio callback. r=pehrsons
On an AudioCallback driver switch, a Fallback driver operates to run the graph
during the switch. The fallback driver is stopped when the
`AudioCallbackDriver::StateCallback` reports the started state. A device is
supposed to start right after the start callback but in OSX, bluetooth devices
can take several hundred milliseconds until the first audio callback arrives.

With this change, the fallback driver is kept alive until the first audio
callback. On first audio callback the Fallback driver is being requested to stop
and silence is returned. Fallback driver, in the next iteration, will run the
graph for last time and stop itself.

Differential Revision: https://phabricator.services.mozilla.com/D61616